updated code for preprocessing

+++ src/preprocess/mol2graph.py
@@ -14,7 +14,7 @@ def mol2graph(mol: Chem.Mol):
         atom_features_list.append(atom_to_feature_vector(atom))
         coords.append(conformer.GetAtomPosition(atom.GetIdx()))
         atom_map[idx] = len(coords) - 1
-    x = np.array(atom_features_list, dtype = np.int64)
+    x = np.array(atom_features_list, dtype = np.int32)
 
     # bonds
     num_bond_features = 3  # bond type, bond stereo, is_conjugated
@@ -36,14 +36,14 @@ def mol2graph(mol: Chem.Mol):
             edge_features_list.append(edge_feature)
 
         # data.edge_index: Graph connectivity in COO format with shape [2, num_edges]
-        edge_index = np.array(edges_list, dtype = np.int64).T
+        edge_index = np.array(edges_list, dtype = np.int32).T
 
         # data.edge_attr: Edge feature matrix with shape [num_edges, num_edge_features]
-        edge_attr = np.array(edge_features_list, dtype = np.int64)
+        edge_attr = np.array(edge_features_list, dtype = np.int32)
 
     else:   # mol has no bonds
-        edge_index = np.empty((2, 0), dtype = np.int64)
-        edge_attr = np.empty((0, num_bond_features), dtype = np.int64)
+        edge_index = np.empty((2, 0), dtype = np.int32)
+        edge_attr = np.empty((0, num_bond_features), dtype = np.int32)
 
     graph = dict()
     graph['edge_index'] = edge_index
