+++ src/preprocess/preprocess.py
@@ -67,8 +67,8 @@ def gen_spatial_edge(dm: np.ndarray, spatial_cutoff: float = 5):
     edge_index = [(x, y) for x, y in zip(src, dst)]
     edge_attr = np.array([SPATIAL_EDGE for _ in edge_index])
 
-    edge_index = np.array(edge_index, dtype=np.int64).T
-    edge_attr = np.array(edge_attr, dtype=np.int64)
+    edge_index = np.array(edge_index, dtype=np.int32).T
+    edge_attr = np.array(edge_attr, dtype=np.int32)
     return edge_index, edge_attr
 
 
@@ -78,8 +78,8 @@ def gen_ligpro_edge(dm: np.ndarray, pocket_cutoff: float):
     edge_index = [(x, y + lig_num_atom) for x, y in zip(lig_idx, pro_idx)]
     edge_index += [(y + lig_num_atom, x) for x, y in zip(lig_idx, pro_idx)]
     edge_attr = np.array([SPATIAL_EDGE for _ in edge_index])
-    edge_index = np.array(edge_index, dtype=np.int64).T
-    edge_attr = np.array(edge_attr, dtype=np.int64)
+    edge_index = np.array(edge_index, dtype=np.int32).T
+    edge_attr = np.array(edge_attr, dtype=np.int32)
     return edge_index, edge_attr
 
 
@@ -137,10 +137,10 @@ def gen_graph(ligand: tuple, pocket: tuple, name: str, protein_cutoff: float, po
         comp_ea = np.vstack([comp_ea, lig_pock_ea])
     # comp_dist = np.array([euclidean(comp_coord[i], comp_coord[j]) for i, j in zip(*comp_ei)], dtype=np.float32)
     # TODO: sort edges
-    comp_num_node = np.array([len(lig_feat), len(pro_feat)], dtype=np.int64)
+    comp_num_node = np.array([len(lig_feat), len(pro_feat)], dtype=np.int32)
     comp_num_edge = np.array(
         [lig_ei.T.shape[0], pro_ei.T.shape[0], lig_pock_ei.T.shape[0], lig_sei.T.shape[0], pro_sei.T.shape[0]],
-        dtype=np.int64)
+        dtype=np.int32)
     return comp_coord, comp_feat, comp_ei, comp_ea, comp_num_node, comp_num_edge
 
 
@@ -159,9 +159,9 @@ def load_pk_data(data_path: Path):
 def to_pyg_graph(raw: list, **kwargs):
     comp_coord, comp_feat, comp_ei, comp_ea, comp_num_node, comp_num_edge, rfscore, gbscore, ecif, pk, name = raw
 
-    d = Data(x=torch.from_numpy(comp_feat).to(torch.long), edge_index=torch.from_numpy(comp_ei).to(torch.long), edge_attr=torch.from_numpy(comp_ea).to(torch.long),
+    d = Data(x=torch.from_numpy(comp_feat).to(torch.int32), edge_index=torch.from_numpy(comp_ei).to(torch.int32), edge_attr=torch.from_numpy(comp_ea).to(torch.int32),
              pos=torch.from_numpy(comp_coord).to(torch.float32), y=torch.tensor([pk], dtype=torch.float32), pdbid=name,
-             num_node=torch.from_numpy(comp_num_node).to(torch.long), num_edge=torch.from_numpy(comp_num_edge).to(torch.long),
+             num_node=torch.from_numpy(comp_num_node).to(torch.int32), num_edge=torch.from_numpy(comp_num_edge).to(torch.int32),
              rfscore=torch.from_numpy(rfscore).to(torch.float32), gbscore=torch.from_numpy(gbscore).to(torch.float32),
              ecif=torch.from_numpy(ecif).to(dtype=torch.float32),
              **kwargs)
